requirements.txt
Django==4.2.15
djangorestframework==3.15.2
ibm_db_django==1.5.3.0
ibm_db==3.2.3
pytest-django==4.8.0
pytest==8.3.2
